requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.1.7
django-environ==0.10.0
django-webpush==0.3.5
http-ece==1.1.0
idna==3.4
marshmallow==3.20.1
mysql-connector-python==8.1.0
mysqlclient==2.1.1
packaging==23.1
protobuf==4.21.12
py-vapid==1.9.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3
pywebpush==1.9.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
